+++ CHANGELOG.md
@@ -11,6 +11,13 @@ The REST API versions in the table are just for your information as the values a
 |\>= 10.0.0|3
 |<= 9.0.1|2
 
+<a name="11.5.0"></a>
+## 2019-03-06 Version [11.5.0](https://github.com/gooddata/gooddata-js/compare/v11.4.0...v11.5.0)
+
+### Changed
+
+- `toAfmResultSpec` now converts native subtotals to `afm.nativeTotals` correctly
+
 <a name="11.4.0"></a>
 ## 2019-03-01 Version [11.4.0](https://github.com/gooddata/gooddata-js/compare/v11.3.1...v11.4.0)